Don't ignore non-running instances, rely on filter instead
We now add the instance-state = ["running"] filter to all queries if
instance-state is not explicitly defined beforehand. This should ensure
that all queries return running instances beforehand (saving some
post-processing later too).

As such, I'm removing an artifact of the old code where we were not
doing this, which was basically to skip all running instances.

Additionally, this will allow an operator to include non-running
instances in a filter should this be desired for whatever reason.