feat: assume role provider credentials
This commit updates the awsutil depedency to its latest version, which
supports the capabilty of using dynamic credentials. The storage service
was refactored to make static credentials optional. The credentials pkg
was refactored to support several attribute fields used for the assume
role provider. The host service still enforces the use of static
credentials.